backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
psycopg2-binary
pydantic[email]
pydantic-settings
pytest
httpx